requirements.txt
django-countries==7.6.1
django-filter==24.3
djangorestframework==3.15.2
Django==5.1.1
Markdown==3.7
django-taggit==6.1.0